[#4389] refactoring
