+++ akvo/rsr/spa/app/modules/editor/actions.js
@@ -72,7 +72,7 @@ export const addProgram = (program) => ({ type: 'ADD_PROGRAM', program })
 export const setProjectTitle = (title) => (dispatch) => {
   dispatch({ type: actionTypes.SAVE_FIELDS, fields: { title }, sectionIndex: 1, noSync: true })
 }
-export const setProjectStatus = (publishingStatus) => (dispatch) => {
-  dispatch({ type: actionTypes.SAVE_FIELDS, fields: { publishingStatus }, sectionIndex: 1, noSync: true })
+export const setProjectStatus = (publishingStatus, hasParent) => (dispatch) => {
+  dispatch({ type: actionTypes.SAVE_FIELDS, fields: { publishingStatus, hasParent }, sectionIndex: 1, noSync: true })
 }
 export const setUser = (user) => ({ type: 'SET_USER', user })
